+++ CHANGELOG.md
@@ -4,7 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [2.1.0] - 2121-07-23
+## [2.2.0] - 2021-07-26
+### Changed
+-  Remove unsupported PHP 7.2 version
+-  Move `script` directory under `src`
+-  Adjust `postInstall` Compose script to:
+    -  update the configuration file, if it finds a deprecated `php_cs.dist`
+    -  delete the deprecated `php_cs.cache` cache file.
+
+## [2.1.0] - 2021-07-23
 ### Changed
 -  Bump `php-cs-fixer` to 3.0 version
 -  Adjust deprecations: